pyproject.toml
[project]
name = "mdn-lstm"
version = "0.1.0"
description = "Mixed Density Network with LSTM for sequence prediction"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
readme = "README.md"
license = { text = "All Rights Reserved" }
requires-python = ">=3.10"
dependencies = ["torch>=2.0.0", "numpy>=1.24.0", "pandas>=2.0.0", "pyyaml>=6.0"]

[project.optional-dependencies]
dev = [
    "invoke>=2.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
mdn-lstm = "mdn_lstm.cli:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[[tool.pdm.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
include_packages = ["torch"]

[tool.pdm.dev-dependencies]
dev = [
    "invoke>=2.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = [
    "N806",
    "N803",
] # Allow uppercase variable/argument names (common in ML: X, y)

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
allow_untyped_calls = true
allow_untyped_defs = true
disable_error_code = ["index", "assignment", "operator", "arg-type"]
